refactor: complete removal of personalityManager facade
- Remove personalityManager.js facade and all related test files
- Update all imports to use core/personality module directly
- Fix index.js to use correct module imports and initialization methods
- Remove personalityManagerCompat.js compatibility wrapper
- Update test file that was still referencing old module path
- Run linter fixes on affected files

All functionality preserved through core/personality module which has
comprehensive test coverage (109 tests, 81.39% coverage).
