+++ src/commands/handlers/status.js
@@ -6,7 +6,7 @@ const { EmbedBuilder } = require('discord.js');
 const logger = require('../../logger');
 const validator = require('../utils/commandValidator');
 const auth = require('../../auth');
-const { listPersonalitiesForUser } = require('../../personalityManager');
+const { listPersonalitiesForUser } = require('../../core/personality');
 const { isAutoResponseEnabled } = require('./autorespond');
 const { getAllActivatedChannels } = require('../../core/conversation');
 const { botPrefix } = require('../../../config');
