--- src/personalityManager.js
@@ -1,282 +1,0 @@
-/**
- * PersonalityManager Facade
- *
- * This file provides backward compatibility for the old personalityManager API
- * while delegating to the new modular implementation in core/personality/
- */
-
-const personalityManager = require('./core/personality');
-const logger = require('./logger');
-
-// Log that we're using the new implementation
-logger.info('[PersonalityManagerFacade] Using new modular personality implementation');
-
-/**
- * Get the maximum word count among all aliases
- * @returns {number} The maximum word count
- */
-function getMaxAliasWordCount() {
-  // Call the method from the PersonalityManager instance
-  if (personalityManager && personalityManager.getMaxAliasWordCount) {
-    const count = personalityManager.getMaxAliasWordCount();
-    logger.debug(`[PersonalityManager] getMaxAliasWordCount returning: ${count}`);
-    return count || 1;
-  }
-  logger.warn('[PersonalityManager] getMaxAliasWordCount method not available, returning default: 1');
-  return 1;
-}
-
-/**
- * Initialize the personality manager (backward compatibility wrapper)
- */
-async function initPersonalityManager(deferOwnerPersonalities = true, options = {}) {
-  return await personalityManager.initialize(deferOwnerPersonalities, options);
-}
-
-/**
- * Register a new personality (backward compatibility wrapper)
- */
-async function registerPersonality(userIdOrFullName, fullNameOrData, dataOrFetchInfo, fetchInfo) {
-  let userId, fullName, data, shouldFetchInfo;
-
-  // Handle old API: registerPersonality(userId, fullName, data, fetchInfo)
-  if (arguments.length >= 3 && typeof fullNameOrData === 'string') {
-    userId = userIdOrFullName;
-    fullName = fullNameOrData;
-    data = dataOrFetchInfo || {};
-    shouldFetchInfo = fetchInfo !== false; // Default true unless explicitly false
-  }
-  // Handle new API: registerPersonality(fullName, userId, activatedChannels)
-  else {
-    fullName = userIdOrFullName;
-    userId = fullNameOrData;
-    data = { activatedChannels: dataOrFetchInfo || [] };
-    shouldFetchInfo = true; // Always fetch in new API
-  }
-
-  // Ensure activatedChannels is an array
-  if (data.activatedChannels) {
-    if (typeof data.activatedChannels === 'string') {
-      data.activatedChannels = [data.activatedChannels];
-    } else if (!Array.isArray(data.activatedChannels)) {
-      data.activatedChannels = [];
-    }
-  }
-
-  // Add fetchInfo flag to data
-  data.fetchInfo = shouldFetchInfo;
-
-  const result = await personalityManager.registerPersonality(fullName, userId, data);
-
-  if (!result.success) {
-    throw new Error(result.error);
-  }
-
-  return personalityManager.getPersonality(fullName);
-}
-
-/**
- * Get a personality by name
- */
-function getPersonality(fullName) {
-  return personalityManager.getPersonality(fullName);
-}
-
-/**
- * Set a personality alias (backward compatibility wrapper)
- */
-async function setPersonalityAlias(alias, fullName, skipSave = false, isDisplayName = false) {
-  // Trim the alias to remove any leading/trailing spaces
-  alias = alias.trim();
-
-  // Check if the alias already exists
-  const existingPersonality = personalityManager.getPersonalityByAlias(alias.toLowerCase());
-  if (existingPersonality && existingPersonality.fullName === fullName) {
-    // Alias already points to this personality, no need to set it again
-    logger.info(`Alias ${alias} already points to ${fullName} - no changes needed`);
-    return { success: true };
-  }
-
-  // If isDisplayName is true and alias already exists for a different personality, create alternate aliases
-  if (isDisplayName && personalityManager.personalityAliases.has(alias.toLowerCase())) {
-    const alternateAliases = [];
-    let alternateAlias = alias;
-
-    // Try to create a smarter alias by using parts of the full personality name
-    const nameParts = fullName.split('-');
-    const aliasParts = alias.split('-');
-
-    // If the personality name has more parts than the alias, try adding the next part
-    if (nameParts.length > aliasParts.length) {
-      // Find which part of the name corresponds to the alias
-      let matchIndex = -1;
-      for (let i = 0; i < nameParts.length; i++) {
-        if (nameParts[i].toLowerCase() === aliasParts[0].toLowerCase()) {
-          matchIndex = i;
-          break;
-        }
-      }
-
-      // If we found a match and there's a next part, use it
-      if (matchIndex >= 0 && matchIndex + 1 < nameParts.length) {
-        alternateAlias = `${alias}-${nameParts[matchIndex + 1]}`;
-      }
-    }
-
-    // If the smart alias is still taken or we couldn't create one, fall back to random
-    if (
-      alternateAlias === alias ||
-      personalityManager.personalityAliases.has(alternateAlias.toLowerCase())
-    ) {
-      // Generate a random suffix with only lowercase letters to match test expectations
-      const chars = 'abcdefghijklmnopqrstuvwxyz';
-      let randomSuffix = '';
-      for (let i = 0; i < 6; i++) {
-        randomSuffix += chars.charAt(Math.floor(Math.random() * chars.length));
-      }
-      alternateAlias = `${alias}-${randomSuffix}`;
-    }
-
-    // Set the alternate alias
-    const result = await personalityManager.setPersonalityAlias(alternateAlias, fullName, skipSave);
-    if (!result.success) {
-      logger.error(`[PersonalityManagerFacade] Failed to set alternate alias: ${result.error}`);
-      return false;
-    }
-
-    alternateAliases.push(alternateAlias);
-
-    // Return object format for backward compatibility
-    return {
-      success: true,
-      alternateAliases,
-    };
-  }
-
-  // Normal alias setting
-  const result = await personalityManager.setPersonalityAlias(alias, fullName, skipSave);
-
-  if (!result.success) {
-    logger.error(`[PersonalityManagerFacade] Failed to set alias: ${result.error}`);
-    // Return object format for tests that expect it
-    if (skipSave || isDisplayName) {
-      return { success: false, error: result.error };
-    }
-    return false;
-  }
-
-  // Return object format for tests that expect it
-  if (skipSave || isDisplayName) {
-    return { success: true };
-  }
-  return true;
-}
-
-/**
- * Get a personality by alias (backward compatibility wrapper)
- */
-function getPersonalityByAlias(personalityOrAlias, alias) {
-  // Handle old API where first param could be null or personality object
-  if (alias !== undefined) {
-    // Old style call: getPersonalityByAlias(null, 'alias')
-    return personalityManager.getPersonalityByAlias(alias);
-  }
-  // New style call: getPersonalityByAlias('alias')
-  return personalityManager.getPersonalityByAlias(personalityOrAlias);
-}
-
-/**
- * Remove a personality (backward compatibility wrapper)
- */
-async function removePersonality(userIdOrFullName, fullNameOrUserId) {
-  let fullName, requestingUserId;
-
-  // Handle old API: removePersonality(userId, fullName)
-  if (arguments.length === 2) {
-    requestingUserId = userIdOrFullName;
-    fullName = fullNameOrUserId;
-  } else {
-    // Handle new API: removePersonality(fullName, userId)
-    fullName = userIdOrFullName;
-    requestingUserId = fullNameOrUserId || 'unknown-user';
-  }
-
-  const result = await personalityManager.removePersonality(fullName, requestingUserId);
-
-  if (!result.success) {
-    // Old API returned false on failure instead of throwing
-    if (
-      result.error === 'Personality not found' ||
-      result.error === 'You can only remove personalities you added'
-    ) {
-      return false;
-    }
-    throw new Error(result.error);
-  }
-
-  return true;
-}
-
-/**
- * List personalities for a user (backward compatibility wrapper)
- */
-function listPersonalitiesForUser(userId) {
-  // If no userId provided, return all personalities (old behavior)
-  if (!userId) {
-    return personalityManager.getAllPersonalities();
-  }
-  return personalityManager.listPersonalitiesForUser(userId);
-}
-
-/**
- * Save all personalities (backward compatibility wrapper)
- */
-async function saveAllPersonalities() {
-  const saved = await personalityManager.save();
-  if (!saved) {
-    throw new Error('Failed to save personalities');
-  }
-}
-
-/**
- * Seed owner personalities
- */
-async function seedOwnerPersonalities(options) {
-  return personalityManager.seedOwnerPersonalities(options);
-}
-
-// For testing - clear all data
-function clearAllData() {
-  if (personalityManager.registry) {
-    personalityManager.registry.clear();
-  }
-}
-
-// Wrapper for getAllPersonalities
-function getAllPersonalities() {
-  return personalityManager.getAllPersonalities();
-}
-
-// Wrapper for save
-async function save() {
-  return personalityManager.save();
-}
-
-// Export the same API as the old personalityManager
-module.exports = {
-  initPersonalityManager,
-  registerPersonality,
-  getPersonality,
-  setPersonalityAlias,
-  getPersonalityByAlias,
-  removePersonality,
-  listPersonalitiesForUser,
-  personalityAliases: personalityManager.personalityAliases,
-  saveAllPersonalities,
-  seedOwnerPersonalities,
-  getAllPersonalities,
-  save,
-  getMaxAliasWordCount,
-  // For backward compatibility with tests
-  personalityData: { clear: clearAllData },
-};
