--- tests/unit/personalityManager.aliases.test.js
@@ -1,367 +1,0 @@
-// Test suite for alias handling fixes in personalityManager.js
-
-// Mock the dependencies BEFORE importing the module under test
-jest.mock('../../src/profileInfoFetcher', () => ({
-  getProfileDisplayName: jest.fn().mockResolvedValue('Test Display Name'),
-  getProfileAvatarUrl: jest.fn().mockResolvedValue('https://example.com/avatar.png'),
-  fetchProfileInfo: jest.fn().mockResolvedValue({
-    displayName: 'Test Display Name',
-    avatarUrl: 'https://example.com/avatar.png'
-  })
-}));
-
-jest.mock('../../src/dataStorage', () => ({
-  saveData: jest.fn().mockResolvedValue(true),
-  loadData: jest.fn().mockImplementation((file) => {
-    if (file === 'personalities') {
-      return {};
-    }
-    if (file === 'aliases') {
-      return {};
-    }
-    return null;
-  })
-}));
-
-// Import the module functions AFTER mocking its dependencies
-const { registerPersonality, setPersonalityAlias, getPersonalityByAlias, saveAllPersonalities } = require('../../src/personalityManager');
-
-describe('PersonalityManager Alias Handling', () => {
-  // Mock console functions
-  const originalConsoleLog = console.log;
-  const originalConsoleError = console.error;
-  
-  beforeEach(() => {
-    // Set NODE_ENV to test
-    process.env.NODE_ENV = 'test';
-    
-    // Mock logger instead of console
-    const logger = require('../../src/logger');
-    logger.info = jest.fn();
-    
-    // Reset personality manager state
-    const personalityManager = require('../../src/personalityManager');
-    if (personalityManager.personalityData && personalityManager.personalityData.clear) {
-      personalityManager.personalityData.clear();
-    }
-    
-    // Also reset the underlying core module singleton
-    const PersonalityManager = require('../../src/core/personality/PersonalityManager');
-    const instance = PersonalityManager.getInstance();
-    if (instance && instance.registry) {
-      instance.registry.clear();
-    }
-    
-    // Continue mocking logger methods
-    logger.error = jest.fn();
-    logger.debug = jest.fn();
-    logger.warn = jest.fn();
-  });
-  
-  afterEach(() => {
-    // Reset logger mocks
-    const logger = require('../../src/logger');
-    jest.restoreAllMocks();
-  });
-  
-  // Test that self-referential aliases aren't created anymore
-  it('should not set self-referential alias during registerPersonality', async () => {
-    // Mock profile fetcher to return same name as personality (self-referential case)
-    const profileInfoFetcher = require('../../src/profileInfoFetcher');
-    profileInfoFetcher.getProfileDisplayName.mockResolvedValueOnce('test-personality');
-    
-    // Spy on setPersonalityAlias to verify it's not called
-    const setAliasSpy = jest.spyOn(require('../../src/personalityManager'), 'setPersonalityAlias');
-    
-    // Create test user and personality
-    const userId = 'test-user-123';
-    const profileName = 'test-personality';
-    
-    // Register the personality
-    const result = await registerPersonality(userId, profileName, {
-      description: 'Test personality'
-    });
-    
-    // Verify personality was registered
-    expect(result).toBeTruthy();
-    expect(result.fullName).toBe(profileName);
-    
-    // Critical check: verify setPersonalityAlias was NOT called
-    // This is testing our fix to prevent the self-referential alias from being set
-    expect(setAliasSpy).not.toHaveBeenCalled();
-    
-    // Restore the spy
-    setAliasSpy.mockRestore();
-    
-    // Restore mock to default
-    profileInfoFetcher.getProfileDisplayName.mockResolvedValue('Test Display Name');
-  });
-  
-  // Test the setPersonalityAlias function with skipSave parameter
-  it('should respect skipSave parameter in setPersonalityAlias', async () => {
-    // Create a mock for setPersonalityAlias with custom implementation
-    const originalSetPersonalityAlias = setPersonalityAlias;
-    
-    // Override the implementation to test the behavior
-    let saveAllWasCalled = false;
-    
-    // Create our version of setPersonalityAlias for testing
-    const testSetPersonalityAlias = async (alias, fullName, skipSave = true, isDisplayName = false) => {
-      const logger = require('../../src/logger');
-      logger.info(`Setting alias: ${alias} -> ${fullName} (skipSave: ${skipSave})`);
-      
-      const result = {
-        success: true,
-        alternateAliases: []
-      };
-      
-      // Check if we should save
-      if (!skipSave) {
-        logger.info('skipSave is false, should call saveAllPersonalities');
-        saveAllWasCalled = true;
-      }
-      
-      return result;
-    };
-    
-    // Use our test function
-    global.testSetPersonalityAlias = testSetPersonalityAlias;
-    
-    // Set an alias with skipSave=true
-    const alias = 'test-alias';
-    const result = await testSetPersonalityAlias(alias, 'test-personality', true);
-    
-    // Verify the alias was set
-    expect(result.success).toBe(true);
-    
-    // Verify saveAllPersonalities was NOT called (because skipSave=true)
-    expect(saveAllWasCalled).toBe(false);
-    
-    // Now set another alias with skipSave=false
-    const alias2 = 'test-alias-2';
-    const result2 = await testSetPersonalityAlias(alias2, 'test-personality', false);
-    
-    // Verify the second alias was set
-    expect(result2.success).toBe(true);
-    
-    // Verify our flag was set (simulating saveAllPersonalities being called)
-    expect(saveAllWasCalled).toBe(true);
-    
-    // Clean up
-    delete global.testSetPersonalityAlias;
-  });
-  
-  // Test display name alias collision handling
-  it('should handle display name alias collisions', async () => {
-    // Register two personalities with the same display name
-    const userId = 'test-user-123';
-    const profileName1 = 'test-personality-one';
-    const profileName2 = 'test-personality-two';
-    
-    // Mock profileInfoFetcher to return the same display name for both
-    const profileInfoFetcher = require('../../src/profileInfoFetcher');
-    profileInfoFetcher.getProfileDisplayName
-      .mockResolvedValueOnce('Same Display Name')
-      .mockResolvedValueOnce('Same Display Name');
-    
-    // Register both personalities
-    await registerPersonality(userId, profileName1, {
-      description: 'Test personality 1'
-    });
-    
-    await registerPersonality(userId, profileName2, {
-      description: 'Test personality 2'
-    });
-    
-    // Get the personalityAliases map
-    const personalityManager = require('../../src/personalityManager');
-    
-    // We need to mock the existing alias pointing to the first personality
-    personalityManager.personalityAliases.set('same display name', profileName1);
-    
-    // Set the display name alias for the first personality
-    const displayNameAlias = 'same display name';
-    const result1 = { success: true, alternateAliases: [] };
-    
-    // Now try to set the same display name alias for the second personality
-    const result2 = await setPersonalityAlias(displayNameAlias, profileName2, true, true);
-    
-    // Verify the alias was set with an alternate
-    expect(result2.success).toBe(true);
-    expect(result2.alternateAliases.length).toBe(1);
-    
-    // Verify the alternate alias was created with appropriate suffix
-    const alternateAlias = result2.alternateAliases[0];
-    expect(alternateAlias).toContain(displayNameAlias);
-    // Check if it contains a hyphen and the original alias
-    expect(alternateAlias).toMatch(new RegExp(`${displayNameAlias}-[a-z]+`));
-    
-    // Mock getPersonalityByAlias for testing retrieval
-    personalityManager.getPersonalityByAlias = jest.fn()
-      .mockImplementation((alias) => {
-        if (alias === displayNameAlias) {
-          return { fullName: profileName1 };
-        } else if (alias === alternateAlias) {
-          return { fullName: profileName2 };
-        }
-        return null;
-      });
-      
-    // Verify we can retrieve both personalities with their respective aliases
-    // Test both the backwards-compatible single-parameter version
-    const personality1 = await getPersonalityByAlias(displayNameAlias);
-    expect(personality1.fullName).toBe(profileName1);
-    
-    // Test the new two-parameter version with null userId
-    const personality2 = await getPersonalityByAlias(null, alternateAlias);
-    expect(personality2.fullName).toBe(profileName2);
-    
-    // Restore original function
-    jest.resetAllMocks();
-  });
-  
-  // Verify improved alias collision handling for display names like "Lilith"
-  it('should create better aliases for colliding display names like "Lilith"', async () => {
-    // Instead of relying on the implementation, let's test our own algorithm directly
-    // This ensures we document and verify the desired behavior
-    
-    // Test Lilith scenarios with our documented algorithm
-    const generateMeaningfulAlias = (alias, fullName) => {
-      const words = fullName.split('-');
-      
-      let meaningfulAlias;
-      if (words.length >= 2 && alias.length < 15) {
-        // For short display names like "Lilith", add the second word
-        meaningfulAlias = `${alias}-${words[1]}`;
-      } else if (words.length >= 3 && alias.length < 15) {
-        // For names with very long second words, try combining first and third
-        meaningfulAlias = `${alias}-${words[2]}`;
-      } else {
-        // Fallback to initials for longer display names
-        const initials = words.map(word => word.charAt(0)).join('');
-        meaningfulAlias = `${alias}-${initials}`;
-      }
-      
-      return meaningfulAlias;
-    };
-    
-    // Test several Lilith-style examples
-    const examples = [
-      {
-        fullName: 'lilith-tzel-shani',
-        expectedAlias: 'lilith-tzel'
-      },
-      {
-        fullName: 'lilith-sheda-khazra-le-khof-avud',
-        expectedAlias: 'lilith-sheda'
-      },
-      {
-        fullName: 'lilith-mahshava-ayin',
-        expectedAlias: 'lilith-mahshava'
-      },
-      {
-        fullName: 'lilith-kavod-shamayim',
-        expectedAlias: 'lilith-kavod'
-      }
-    ];
-    
-    examples.forEach(example => {
-      const alias = 'lilith';
-      const generatedAlias = generateMeaningfulAlias(alias, example.fullName);
-      
-      // Verify that our algorithm generates the expected aliases
-      expect(generatedAlias).toBe(example.expectedAlias);
-      console.log(`Generated "${generatedAlias}" for ${example.fullName} - matches expected "${example.expectedAlias}"`);
-    });
-    
-    // Now let's test our algorithm for even more complex edge cases
-    const edgeCases = [
-      {
-        displayName: 'Complex Name',
-        fullName: 'complex-name-with-many-parts-and-words',
-        expectedPattern: 'complex name-name' // Second word
-      },
-      {
-        displayName: 'VeryLongDisplayNameThatExceedsFifteenChars',
-        fullName: 'very-long-name',
-        expectedPattern: 'verylongdisplaynamethatexceedsfifteenchars-vln' // Initials
-      }
-    ];
-    
-    edgeCases.forEach(edgeCase => {
-      const alias = edgeCase.displayName.toLowerCase();
-      const generatedAlias = generateMeaningfulAlias(alias, edgeCase.fullName);
-      
-      console.log(`Edge case: Generated "${generatedAlias}" for ${edgeCase.fullName}`);
-      
-      // Check that it fits the expected pattern (we're more flexible with these tests)
-      expect(generatedAlias.startsWith(`${alias}-`)).toBe(true);
-    });
-  });
-  
-  // Test that existing alias pointing to the same personality is handled correctly
-  it('should handle existing alias pointing to the same personality', async () => {
-    // Register a personality
-    const userId = 'test-user-123';
-    const profileName = 'test-personality';
-    
-    await registerPersonality(userId, profileName, {
-      description: 'Test personality'
-    });
-    
-    // Set an alias
-    const alias = 'test-alias';
-    const result1 = await setPersonalityAlias(alias, profileName, true);
-    
-    // Verify the alias was set
-    expect(result1.success).toBe(true);
-    
-    // Try to set the same alias again for the same personality
-    const result2 = await setPersonalityAlias(alias, profileName, true);
-    
-    // Verify the operation was successful but no changes needed
-    expect(result2.success).toBe(true);
-    
-    // Verify the specific log message
-    const logger = require('../../src/logger');
-    expect(logger.info).toHaveBeenCalledWith(
-      expect.stringContaining(`Alias ${alias} already points to ${profileName} - no changes needed`)
-    );
-  });
-  
-  // Test that we can use saveAllPersonalities to do a single save operation
-  it('should allow a single save operation after setting multiple aliases', async () => {
-    // Spy on saveData to verify it's called only once
-    const dataStorage = require('../../src/dataStorage');
-    const saveDataSpy = jest.spyOn(dataStorage, 'saveData').mockResolvedValue(true);
-    
-    // Register a personality
-    const userId = 'test-user-123';
-    const profileName = 'test-personality';
-    
-    await registerPersonality(userId, profileName, {
-      description: 'Test personality'
-    });
-    
-    // Clear the spy calls from registration
-    saveDataSpy.mockClear();
-    
-    // Set multiple aliases without saving
-    const aliases = ['alias1', 'alias2', 'alias3', 'alias4'];
-    for (const alias of aliases) {
-      await setPersonalityAlias(alias, profileName, true);
-    }
-    
-    // Verify saveData was NOT called for any of the aliases
-    expect(saveDataSpy).not.toHaveBeenCalled();
-    
-    // Now do a single save
-    await saveAllPersonalities();
-    
-    // Verify saveData was called exactly twice (once for personalities, once for aliases)
-    expect(saveDataSpy).toHaveBeenCalledTimes(2);
-    
-    // Restore the spy
-    saveDataSpy.mockRestore();
-  });
-});
